mod: 보안코드 상태값도 커스텀 훅으로 분리
