+++ src/pages/AddingCard/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useCardNumber } from './useCardNumber';
 export { default as useExpiredDate } from './useExpiredDate';
 export { default as useCardOwner } from './useCardOwner';
+export { default as useSecretCode } from './useSecretCode';
 export { default as useCompanyPicker } from './useCompanyPicker';
